Move tablemodel point addition/deletion handling from GraphicsScene to Window
